AIRA 1.0/requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
streamlit
langchain-groq
pypdf
faiss-cpu
google-generativeai
chromadb
llama-cpp-python
langchain-ollama
ipykernel
sentence-transformers
openpyxl
streamlit-extras